+++ test/Stormpath.Owin.UnitTest/ConfigurationHelper.cs
@@ -18,7 +18,7 @@ namespace Stormpath.Owin.UnitTest
 
             var integrationConfig = new IntegrationConfiguration(
                 compiledConfig, 
-                new OktaEnvironmentConfiguration("fooAuthServerId", "fooClientId", "fooClientSecret123456"),
+                new OktaEnvironmentConfiguration("fooAuthServerId", new[] { "my-audience" }, "fooClientId", "fooClientSecret123456"),
                 new KeyValuePair<string, ProviderConfiguration>[0],
                 null);